chore: Move optimize_subquery_sort into optimizer as a new rule EliminateSort
