+++ datafusion/sql/src/relation/mod.rs
@@ -19,13 +19,14 @@ use std::sync::Arc;
 
 use crate::planner::{ContextProvider, PlannerContext, SqlToRel};
 
-use datafusion_common::tree_node::{Transformed, TreeNode};
 use datafusion_common::{
     not_impl_err, plan_err, DFSchema, Diagnostic, Result, Span, Spans, TableReference,
 };
 use datafusion_expr::builder::subquery_alias;
 use datafusion_expr::{expr::Unnest, Expr, LogicalPlan, LogicalPlanBuilder};
 use datafusion_expr::{Subquery, SubqueryAlias};
+use datafusion_optimizer::eliminate_sort::EliminateSort;
+use datafusion_optimizer::{OptimizerContext, OptimizerRule};
 use sqlparser::ast::{FunctionArg, FunctionArgExpr, Spanned, TableFactor};
 
 mod join;
@@ -162,7 +163,10 @@ impl<S: ContextProvider> SqlToRel<'_, S> {
             }
         };
 
-        let optimized_plan = optimize_subquery_sort(plan)?.data;
+        let opt_subquery_sort = EliminateSort::new();
+        let optimized_plan = opt_subquery_sort
+            .rewrite(plan, &OptimizerContext::default())?
+            .data;
         if let Some(alias) = alias {
             self.apply_table_alias(optimized_plan, alias)
         } else {
@@ -224,31 +228,4 @@ impl<S: ContextProvider> SqlToRel<'_, S> {
             })),
         }
     }
-}
-
-fn optimize_subquery_sort(plan: LogicalPlan) -> Result<Transformed<LogicalPlan>> {
-    // When initializing subqueries, we examine sort options since they might be unnecessary.
-    // They are only important if the subquery result is affected by the ORDER BY statement,
-    // which can happen when we have:
-    // 1. DISTINCT ON / ARRAY_AGG ... => Handled by an `Aggregate` and its requirements.
-    // 2. RANK / ROW_NUMBER ... => Handled by a `WindowAggr` and its requirements.
-    // 3. LIMIT => Handled by a `Sort`, so we need to search for it.
-    let mut has_limit = false;
-    let new_plan = plan.transform_down(|c| {
-        if let LogicalPlan::Limit(_) = c {
-            has_limit = true;
-            return Ok(Transformed::no(c));
-        }
-        match c {
-            LogicalPlan::Sort(s) => {
-                if !has_limit {
-                    has_limit = false;
-                    return Ok(Transformed::yes(s.input.as_ref().clone()));
-                }
-                Ok(Transformed::no(LogicalPlan::Sort(s)))
-            }
-            _ => Ok(Transformed::no(c)),
-        }
-    });
-    new_plan
 }
